% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedModify.R
\name{.check_parents_before_children}
\alias{.check_parents_before_children}
\title{Check pedigree order}
\usage{
.check_parents_before_children(x)
}
\arguments{
\item{x}{a \code{\link{linkdat}} object}
}
\description{
Checks a pedigree or linkdat object if parents come before their children
}
\examples{
x1 = nuclearPed(1)
x2 = linkdat(x1$pedigree[3:1, ])
stopifnot(.parents_before_children(x1), !.parents_before_children(x2))

}
