% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramlink.R
\docType{package}
\name{paramlink-package}
\alias{paramlink}
\alias{paramlink-package}
\title{paramlink: Parametric Linkage and Other Pedigree Analysis in R}
\description{
A suite of tools for analysing pedigrees with marker data, including parametric linkage analysis, forensic computations, relatedness analysis and marker simulations. The core of the package is an implementation of the Elston-Stewart algorithm for pedigree likelihoods, extended to allow mutations as well as complex inbreeding. Features for linkage analysis include singlepoint LOD scores, power analysis, and multipoint analysis (the latter through a wrapper to the 'MERLIN' software). Forensic applications include exclusion probabilities, genotype distributions and conditional simulations. Data from the 'Familias' software can be imported and analysed in 'paramlink'. Finally, 'paramlink' offers many utility functions for creating, manipulating and plotting pedigrees with or without marker data (the actual plotting is done by the 'kinship2' package).
}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no}

Other contributors:
\itemize{
  \item Thore Egeland [contributor]
  \item Guro Doerum [contributor]
}

}
\keyword{internal}
