% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDatfile.R
\name{readDatfile}
\alias{readDatfile}
\title{Read dat file in LINKAGE format}
\usage{
readDatfile(datfile, chrom, comment_string = "<<", write_to = NULL)
}
\arguments{
\item{datfile}{character. The path to the dat file.}

\item{chrom}{integer chromosome number (needed to create the MERLIN map).}

\item{comment_string}{character indicating comments (which are removed before
processing).}

\item{write_to}{a character prefix used for naming the output files, or NULL
if no files should be written.}
}
\value{
If \code{write_to} is NULL, a list of data.frames named \code{dat},
  \code{map} and \code{freq}.
}
\description{
Converts dat files in LINKAGE format to dat/map/freq files in MERLIN format
}
\examples{

# No example given.

}
