% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{toyped}
\alias{toyped}
\title{Toy pedigree for linkage analysis}
\format{
A data frame with 4 rows and 8 columns
}
\usage{
toyped
}
\description{
Toy pedigree with 4 individuals typed at a single SNP marker. Individual 1 is
missing one allele.
}
\details{
The format is standard LINKAGE (pre-makeped) format, with columns as follows:
\itemize{
  \item FAMID. Family ID
  \item ID. Individual ID
  \item FID. Father ID
  \item MID. Mother ID
  \item SEX. Gender (male=1, female=2)
  \item AFF. Affection status (unaffected=1, affected=2, unkonwn=0)
  \item M_A1. First allele of marker 1
  \item M_A2. Second allele of marker 1
}
}
\examples{

toyped
linkdat(toyped)

}
\keyword{datasets}
