% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{twoloops}
\alias{twoloops}
\title{A consanguineous pedigree}
\format{A data frame with 17 rows and 6 columns. 
See \code{\link{toyped}} for details about the format.}
\usage{
twoloops
}
\description{
A consanguineous pedigree with two inbreeding loops.
}
\examples{

x = linkdat(twoloops)
plot(x)

}
\keyword{datasets}
